DONE -  Added judgement of request type, fix a bug that system will not alert when nickname conflict.

+++ php/functions/delete-menu.php
@@ -1,8 +1,9 @@
 <?php
 include('../common/config.php');
+if ($_SERVER["REQUEST_METHOD"] == "POST") {
+  $date = $_POST['date'];
+  $sql_del_menu = "DELETE FROM `lunch_menu` WHERE `date` = '$date'";
+  $result = mysqli_query($mysqlConnection, $sql_del_menu);
 
-$date = $_POST['date'];
-$sql_del_menu = "DELETE FROM `lunch_menu` WHERE `date` = '$date'";
-$result = mysqli_query($mysqlConnection, $sql_del_menu);
-
-echo(json_encode($result));
+  echo(json_encode($result));
+}
